refactor: improve and refine the usage scope of interfaces and classes
